Use native document windows
